+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - use "to_lowercase" for profile name checking
 
+### Changed
+
+- "bad_request" error to "duplicate" error
+
 ## [0.2.2]
 
 ### Added